requirements.txt
fastapi
uvicorn
redis
aiohttp
pydantic>=2.0
pydantic-settings
sse-starlette
httpx
scikit-learn
numpy